Modified grunt file to support 1 conf.js passing suite, chromeo options for debug, and path to config in params.
Removed unnecesary conf.js files and used the args in grunt file to control the differences.

Renamed base.conf.js to just conf.js.

Refactored base.conf into conf.js  and control the different settings through protractor configs instead of extra code. Added accepting of params from cli, set default chrome arg to headless and is overridden by the cli args in grunt for debug
