--- tests/e2e.tests.debug.conf.js
@@ -1,2 +1,0 @@
-const BaseConfig = require('./base.conf.js');
-module.exports = new BaseConfig('e2e', { headless: false });
