--- tests/performance.tests-only.conf.js
@@ -1,3 +1,0 @@
-const BaseConfig = require('./base.conf.js');
-module.exports = new BaseConfig('performance');
-
